requirements.txt
fastapi
uvicorn
httpx
python-dotenv
mysql-connector-python